# subproyectoIntegracion-l8-2/src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: 'https://city-events-api.ew.r.appspot.com/api'
  - url: 'http://localhost:8089/api'
info:
  description: Api about cities and events that take place in them
  version: 1.0.0
  title: CityEventsAPI
  contact:
    email: javnunrui@alum.us.es
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /cities:
    get:
      tags:
        - cities
      summary: Get cities
      description: Get all cities with the right attributes
      operationId: getCities
      parameters:
        - name: limit
          in: query
          description: The maximum number of items to return
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: q
          in: query
          description: Query to filter
          required: false
          schema:
            type: string
        - name: name
          in: query
          description: Filter cities by name
          required: false
          schema:
            type: string
        - name: isEmpty
          in: query
          description: Filter cities with no events
          required: false
          schema:
            type: boolean
        - name: fields
          in: query
          description: Returns only the properties of the cities
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: Order results by name
          required: false
          schema:
            type: string
            enum:
              - name
              - '-name'
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/City'
        '400':
          description: Bad request
    post:
      tags:
        - cities
      summary: Create a city
      description: Create a city with the right attributes
      operationId: createCity
      requestBody:
        description: City to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/City'
      responses:
        '201':
          description: City created correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '400':
          description: Bad request
    put:
      tags:
        - cities
      summary: Update a city
      description: Update a city introducing new values for certain attributes
      operationId: updateCity
      requestBody:
        description: City to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/City'
      responses:
        '204':
          description: No content. City updated
        '400':
          description: Bad request
        '404':
          description: City not found
  '/cities/{id}':
    get:
      tags:
        - cities
      summary: Get a city by ID
      description: Get a city by ID with the right attributes
      operationId: getCity
      parameters:
        - name: id
          in: path
          description: Id of the city
          required: true
          schema:
            type: string
      responses:
        '200':
          description: City with corresponding id found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '404':
          description: City not found
    delete:
      tags:
        - cities
      summary: Delete a city
      description: Delete a city
      operationId: deleteCity
      parameters:
        - name: id
          in: path
          description: Id of the city
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content. City deleted
        '404':
          description: City not found
  '/cities/totalPrice/{id}':
    get:
      tags:
        - cities
      summary: Get total price of all events of a city by ID
      description: Get the minimum and maximum total price of all events of a city by ID
      operationId: getTotalPriceCity
      parameters:
        - name: id
          in: path
          description: Id of the city
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Calculated total price for city with corresponding id
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: City not found
  '/cities/mostRelevantOrganizer/{id}':
    get:
      tags:
        - cities
      summary: Gets the organizer who has held the most events in a city by ID
      description: Gets the organizer who has held the most events in a city by ID along with the number of these
      operationId: getMostRelevantOrganizerCity
      parameters:
        - name: id
          in: path
          description: Id of the city
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Organizer and number of events obtained for city with corresponding id
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: City not found
  '/cities/eventCategories/{id}':
    get:
      tags:
        - cities
      summary: Get list of event numbers by category of city by ID
      description: Get list of event numbers by category of city by ID
      operationId: getEventCategoriesCity
      parameters:
        - name: id
          in: path
          description: Id of the city
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of event numbers by category obtained for city with corresponding id
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: City not found
  '/cities/{cityId}/{eventId}':
    post:
      tags:
        - cities
      summary: Add event to city
      description: Add event with eventId to city with cityId
      operationId: addEventToCity
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters:
        - name: cityId
          in: path
          description: Id of the city
          required: true
          schema:
            type: string
        - name: eventId
          in: path
          description: Id of the event
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Event added to city correctly
          content:
            application/json:
              schema:
                required:
                  - name
                  - description
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                  events:
                    type: array
                    items:
                      type: object
                      required:
                        - name
                        - description
                        - organizer
                        - category
                        - location
                        - date
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
                        organizer:
                          type: string
                        category:
                          type: string
                        location:
                          type: string
                        date:
                          type: string
                        price:
                          type: string
                        fqas:
                          type: array
                          items:
                            type: object
                            required:
                              - question
                              - answer
                example:
                  id: c9
                  name: Barcelona
                  description: The capital of Catalonia in Spain...
                  events:
                    - id: e5
                      name: Iron Maiden Concert
                      description: Iron Maiden returns to Spain this summer
                      organizer: Delegacion de arte de Catalunya
                      category: Music
                      location: Estadi Olimpic Lluis Companys
                      date: '29/06/2022'
                      price: '100-200 euros'
                      fqas:
                        - id: f15
                          question: Area for reduced mobility?
                          answer: There is a conveniently equipped area for people with reduced mobility...
                    - id: e6
                      name: Mobile World Congress
                      description: Great fair dedicated to mobile telephony and other technological devices...
                      organizer: GSM Association
                      category: Technology
                      location: Fira Gran Via
                      date: '28/02/2022'
                      price: '700-4200 euros'
                      fqas: null
        '400':
          description: Bad request
        '404':
          description: City or event not found
    delete:
      tags:
        - cities
      summary: Delete event from city
      description: Delete event from city
      operationId: deleteEventFromCity
      parameters:
        - name: cityId
          in: path
          description: Id of the city
          required: true
          schema:
            type: string
        - name: eventId
          in: path
          description: Id of the event
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content. Event deleted from city
        '404':
          description: City or event not found
  /events:
    get:
      tags:
        - events
      summary: Get events
      description: Get all events with the right attributes
      operationId: getEvents
      parameters:
        - name: limit
          in: query
          description: The maximum number of items to return
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: q
          in: query
          description: Query to filter
          required: false
          schema:
            type: string
        - name: name
          in: query
          description: Filter events by name
          required: false
          schema:
            type: string
        - name: organizer
          in: query
          description: Filter events by organizer
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Filter events by category
          required: false
          schema:
            type: string
        - name: location
          in: query
          description: Filter events by location
          required: false
          schema:
            type: string
        - name: isEmpty
          in: query
          description: Filter events with no fqas
          required: false
          schema:
            type: boolean
        - name: fields
          in: query
          description: Returns only the properties of the events
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: Order results by category, organizer or date
          required: false
          schema:
            type: string
            enum:
              - category
              - '-category'
              - organizer
              - '-organizer'
              - date
              - '-date'
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Bad request
    post:
      tags:
        - events
      summary: Create a event
      description: Create a event with the right attributes
      operationId: createEvent
      requestBody:
        description: Event to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: Event created correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Bad request
    put:
      tags:
        - events
      summary: Update a event
      description: Update a event introducing new values for certain attributes
      operationId: updateEvent
      requestBody:
        description: Event to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '204':
          description: No content. Event updated
        '400':
          description: Bad request
        '404':
          description: Event not found
  '/events/{id}':
    get:
      tags:
        - events
      summary: Get a event by ID
      description: Get a event by ID with the right attributes
      operationId: getEvent
      parameters:
        - name: id
          in: path
          description: Id of the event
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event with corresponding id found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
    delete:
      tags:
        - events
      summary: Delete a event
      description: Delete a event
      operationId: deleteEvent
      parameters:
        - name: id
          in: path
          description: Id of the event
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content. Event deleted
        '404':
          description: Event not found
...
components:
  schemas:
    City:
      type: object
      required:
        - name
        - description
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
      example:
        id: c9
        name: Barcelona
        description: The capital of Catalonia in Spain...
        events:
          - id: e5
            name: Iron Maiden Concert
            description: Iron Maiden returns to Spain this summer
            organizer: Delegacion de arte de Catalunya
            category: Music
            location: Estadi Olimpic Lluis Companys
            date: '29/06/2022'
            price: '100-200 euros'
            fqas:
              - id: f15
                question: Area for reduced mobility?
                answer: There is a conveniently equipped area for people with reduced mobility...
          - id: e6
            name: Mobile World Congress
            description: Great fair dedicated to mobile telephony and other technological devices...
            organizer: GSM Association
            category: Technology
            location: Fira Gran Via
            date: '28/02/2022'
            price: '700-4200 euros'
            fqas: null
    Event:
      type: object
      required:
        - name
        - description
        - organizer
        - category
        - location
        - date
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        organizer:
          type: string
        category:
          type: string
        location:
          type: string
        date:
          type: string
        price:
          type: string
        fqas:
          type: array
          items:
            $ref: '#/components/schemas/Fqa'
      example:
        id: e5
        name: Iron Maiden Concert
        description: Iron Maiden returns to Spain this summer
        organizer: Delegacion de arte de Catalunya
        category: Music
        location: Estadi Olimpic Lluis Companys
        date: '29/06/2022'
        price: '100-200 euros'
        fqas:
          - id: f15
            question: Area for reduced mobility?
            answer: There is a conveniently equipped area for people with reduced mobility...
    Fqa:
      type: object
      required:
        - question
        - answer
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
      example:
        id: f15
        question: Area for reduced mobility?
        answer: There is a conveniently equipped area for people with reduced mobility...